refactor(python)!: drop eager re-exports from bloqade.lanes __init__
bloqade layering runs bloqade.gemini on top of bloqade.lanes, but lanes
submodules import bloqade.gemini at module load (e.g. rewrite.circuit2place
-> bloqade.gemini.logical). An eager re-export in bloqade.lanes/__init__
re-enters an in-progress bloqade.gemini import and breaks initialization
(surfacing once a dialect imports LocationAddress from
bloqade.lanes.bytecode.encoding). Empty the top level and require consumers
to import from submodules / bloqade.gemini.

Migrate the affected demos and the Steane regression test to the submodule
imports. Broader top-level cleanup tracked in #801.

BREAKING CHANGE: bloqade.lanes no longer re-exports device classes, Metrics,
noise-model helpers, or steane7_* at the package root. Import them from their
submodules (bloqade.lanes.metrics, bloqade.lanes.noise_model,
bloqade.lanes.steane_defaults) or from bloqade.gemini.
