+++ demo/simulator_device_demo.py
@@ -8,7 +8,7 @@ from kirin.dialects import ilist
 from bloqade import qubit, squin
 from bloqade.gemini import logical
 from bloqade.gemini.logical.stdlib import default_post_processing
-from bloqade.lanes import GeminiLogicalSimulator
+from bloqade.gemini import GeminiLogicalSimulator
 
 
 # helper functions to analyze statistical distribution of logical measurements
